add: consumer with dummy message
